feat: converted sandbox example to e2e
