+++ docs/docs/dev_docs/getting_started/sandbox.md
@@ -171,9 +171,13 @@ Add a `tsconfig.json` file into the project root, here is an example:
 yarn add @aztec/aztec.js @aztec/noir-contracts
 ```
 
-7. Create an `index.ts` file in the `src` directory and add the following snippet
+7. Create an `index.ts` file in the `src` directory and add the following imports:
 
-#include_code index /docs/src/code_examples/sandbox_example.ts typescript
+#include_code imports /yarn-project/end-to-end/src/e2e_sandbox_example.ts typescript
+
+and the following setup code:
+
+#include_code setup /yarn-project/end-to-end/src/e2e_sandbox_example.ts typescript
 
 8. Finally, run the package:
 
@@ -198,7 +202,7 @@ The next step is to create some accounts. An in-depth explaining about accounts 
 
 Continue with adding the following to the `index.ts` file in our example:
 
-#include_code Accounts /docs/src/code_examples/sandbox_example.ts typescript
+#include_code Accounts /yarn-project/end-to-end/src/e2e_sandbox_example.ts typescript
 
 Running `yarn start` should now output:
 
@@ -224,7 +228,7 @@ If you were looking at your terminal that is running the Sandbox you should hope
 
 Now that we have our accounts setup, let's move on to deploy our private token contract. Add this to `index.ts`:
 
-#include_code Deployment /docs/src/code_examples/sandbox_example.ts typescript
+#include_code Deployment /yarn-project/end-to-end/src/e2e_sandbox_example.ts typescript
 
 `yarn start` will now give the following output:
 
@@ -252,7 +256,7 @@ The Private Token Contract emits an unencrypted log message during construction:
 
 We can retrieve this emitted log using the `getUnencryptedLogs()` api:
 
-#include_code Logs /docs/src/code_examples/sandbox_example.ts typescript
+#include_code Logs /yarn-project/end-to-end/src/e2e_sandbox_example.ts typescript
 
 Our output will now be:
 
@@ -278,7 +282,7 @@ A token contract wouldn't be very useful if you aren't able to query the balance
 
 Call this function using the following code:
 
-#include_code Balance /docs/src/code_examples/sandbox_example.ts typescript
+#include_code Balance /yarn-project/end-to-end/src/e2e_sandbox_example.ts typescript
 
 Running now should yield output:
 
@@ -312,7 +316,7 @@ Now lets transfer some funds from Alice to Bob by calling the `transfer` functio
 
 We will again view the unencrypted logs emitted by the function and check the balances after the transfer:
 
-#include_code Transfer /docs/src/code_examples/sandbox_example.ts typescript
+#include_code Transfer /yarn-project/end-to-end/src/e2e_sandbox_example.ts typescript
 
 Our output should now look like this:
 
@@ -346,7 +350,7 @@ Finally, the contract has a `mint` function that can be used to generate new tok
 
 Let's mint some tokens to Bob's account:
 
-#include_code Mint /docs/src/code_examples/sandbox_example.ts typescript
+#include_code Mint /yarn-project/end-to-end/src/e2e_sandbox_example.ts typescript
 
 Our complete output should now be:
